[errors]

+++ micro-metadb/service/host.go
@@ -566,7 +566,7 @@ func (handler *DBServiceHandler) UpdateHostStatus(ctx context.Context, in *dbpb.
 		// return nil to use rsp
 		return nil
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 	return nil
 }
 
@@ -591,7 +591,7 @@ func (handler *DBServiceHandler) ReserveHost(ctx context.Context, in *dbpb.DBRes
 		// return nil to use rsp
 		return nil
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 	return nil
 }
 
@@ -721,14 +721,14 @@ func (handler *DBServiceHandler) GetHierarchy(ctx context.Context, in *dbpb.DBGe
 	}
 	wholeTree := handler.buildHierarchy(Items)
 	if wholeTree == nil {
-		out.Rs.Code = common.TIEM_RESOURCE_NO_STOCK
+		out.Rs.Code = int32(common.TIEM_RESOURCE_NO_STOCK)
 		out.Rs.Message = fmt.Sprintf("no stocks with filter:%v", filter)
 		return nil
 	}
 	root := handler.trimTree(wholeTree, resource.FailureDomain(in.Level), int(in.Depth))
 	copyHierarchyToRsp(root, &out.Root)
 
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 
 	return nil
 }
@@ -795,7 +795,7 @@ func (handler *DBServiceHandler) GetStocks(ctx context.Context, in *dbpb.DBGetSt
 		out.Stocks.FreeDiskCount += int32(stock.FreeDiskCount)
 		out.Stocks.FreeDiskCapacity += int32(stock.FreeDiskCapacity)
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 
 	return nil
 }
